Persistent account  Changes to be committed: 	modified:   app/Http/Controllers/UserController.php 	modified:   config/app.php 	modified:   database/seeds/UsersTableSeeder.php 	modified:   resources/views/users/index.blade.php

+++ app/Http/Controllers/UserController.php
@@ -128,6 +128,10 @@ class UserController extends Controller
     public function destroy(User $user)
     {
         //
+        return config('app.super_user');
+        if($user->email==config('app.super_user')){
+            return back();
+        }
         $user->delete();
         return back();
     }
